fixed dbreader

+++ car-sharing-system/car-sharing-system/Models/Car.cs
@@ -13,14 +13,14 @@ namespace car_sharing_system.Models
     public String vehicleType { get; set; }
     public int seats { get; set; }
     public double rate { get; set; }
-    public double coorX { get; set; }
-    public double coorY { get; set; }
+    public decimal coorX { get; set; }
+    public decimal coorY { get; set; }
     public char transmission { get; set; }
     public int tankSize { get; set; }
     public double fuelConsumption { get; set; }
     public int avgRange { get; set; }
 
-    public Car(string numberPlate, string brand, string model, string vehicleType, int seats, double rate, double coorX, double coorY)
+    public Car(string numberPlate, string brand, string model, string vehicleType, int seats, double rate, decimal coorX, decimal coorY)
     {
       this.numberPlate = numberPlate;
       this.brand = brand;
